+++ Mandelbrot.Domain.Calculation/Algorithms/MandelbrotAlgorithm.cs
@@ -4,7 +4,7 @@
 	{
 		private const int CutOff = (1 << 16);
 
-		public override Rectangle<decimal> DefaultScale => new Rectangle<decimal>(-2.5m, -1, 1, 1);
+		public override Rectangle<double> DefaultScale => new Rectangle<double>(-2.5d, -1, 1, 1);
 
 		public override double CalculateSinglePoint(int x, int y, IFractalSettings settings)
 		{
@@ -17,10 +17,10 @@
 
 		protected override string Name => "Mandelbrot Set";
 
-		private static int RunMandelbrotLoop(decimal scaledX, decimal scaledY, int maxIterations)
+		private static int RunMandelbrotLoop(double scaledX, double scaledY, int maxIterations)
 		{
-			decimal currentX = 0;
-			decimal currentY = 0;
+            double currentX = 0;
+            double currentY = 0;
 
 			int iteration;
 			for (iteration = 0; iteration < maxIterations - 1; iteration++)
