+++ Mandelbrot.Domain.Rendering.Interface/Output/ScreenExtension.cs
@@ -2,9 +2,9 @@
 {
 	public static class ScreenExtension
 	{
-		public static Rectangle<decimal> ScaleFractalRectToScreenSelection(
+		public static Rectangle<double> ScaleFractalRectToScreenSelection(
 			this IScreen screen,
-			Rectangle<decimal> fractalRect,
+			Rectangle<double> fractalRect,
 			Rectangle<int> screenSelection)
 		{
 			var left = ScaleScreenValue(screenSelection.Left, screen.Width, fractalRect.Left, fractalRect.Right);
@@ -12,10 +12,10 @@
 			var top = ScaleScreenValue(screenSelection.Top, screen.Height, fractalRect.Top, fractalRect.Bottom);
 			var bottom = ScaleScreenValue(screenSelection.Bottom, screen.Height, fractalRect.Top, fractalRect.Bottom);
 
-			return new Rectangle<decimal>(left, top, right, bottom);
+			return new Rectangle<double>(left, top, right, bottom);
 		}
 
-		private static decimal ScaleScreenValue(decimal screenValue, decimal screenSize, decimal fractalStart, decimal fractalEnd)
+		private static double ScaleScreenValue(double screenValue, double screenSize, double fractalStart, double fractalEnd)
 		{
 			var screenPercentage = screenValue / screenSize;
 			var scaleSize = fractalEnd - fractalStart;
